Fix RESET pin

+++ Firmware/DAP.h
@@ -194,8 +194,8 @@ extern void     JTAG_Sequence   (uint32_t info,  uint8_t *tdi, uint8_t *tdo);
 extern void     JTAG_IR         (uint32_t ir);
 extern uint32_t JTAG_ReadIDCode (void);
 extern void     JTAG_WriteAbort (uint32_t data);
-extern uint8_t  JTAG_Transfer   (uint32_t request, uint32_t *data);
-extern uint8_t  SWD_Transfer    (uint32_t request, uint32_t *data);
+extern uint8_t  JTAG_Transfer   (uint8_t request, uint32_t *data);
+extern uint8_t  SWD_Transfer    (uint8_t request, uint32_t *data);
 
 extern void     Delayms         (uint32_t delay);
